contents predrawn

+++ src/wc_handler/page/page_json.rs
@@ -16,16 +16,8 @@ impl PageJson {
         PageJson { data: Some(data) }
     }
 
-    // pub fn data(&self) -> Option<&json::JsonValue> {
     pub fn value(&self) -> Option<&json::JsonValue> {
         self.data.as_ref()
-        // match self.data.as_ref() {
-        //     Some(v) => Some(v),
-        //     None => {
-        //         eprintln!("Failed to get data in json");
-        //         None
-        //     }
-        // }
     }
 
     // pub fn value_mut(&mut self) -> Option<&mut json::JsonValue> {
@@ -49,7 +41,7 @@ impl PageJson {
     //     }
     // }
 
-    pub fn rev(&self) -> Option<u32> {
+    pub fn rev(&self) -> Option<usize> {
         // match self.data()?["data"]["page"]["rev"] {
         match self.value()?["data"]["page"]["rev"] {
             // case rev=10: Number(Number { category: 1, exponent: 0, mantissa: 10 })
@@ -65,7 +57,7 @@ impl PageJson {
             // case: rev="12" ( with "" )
             json::JsonValue::Short(short) => {
                 let rev = short.as_str();
-                match u32::from_str(rev) {
+                match usize::from_str(rev) {
                     Ok(v) => return Some(v),
                     Err(_) => {
                         eprintln!("Failed to get rev");
@@ -73,71 +65,17 @@ impl PageJson {
                     }
                 }
             }
-            _ => None::<u32>,
+            _ => None::<usize>,
         };
 
         None
     }
 
     // rev counted up from current rev
-    pub fn rev_uped(&self) -> Option<u32> {
+    pub fn rev_uped(&self) -> Option<usize> {
         let rev = self.rev()?;
         Some(rev + 1)
     }
-
-    // data["data"]["navi"] {
-    // pub fn navi(&self) -> Option<&json::object::Object> {
-    // pub fn navi(&self) -> Option<&json::JsonValue> {
-    // pub fn navi(&self) -> Option<&json::JsonValue::Array> {
-    // pub fn navi(&self) -> Option<&json::Array> {
-    //     // DBG
-    //     // info!("fn navi");
-
-    //     // let data = self.data()?;
-    //     let value = self.value()?;
-
-    //     // DBG
-    //     // info!("fn navi cp1");
-
-    //     if value["data"]["navi"].is_empty() {
-    //         return None;
-    //     }
-
-    //     // DBG
-    //     // info!("fn navi cp2 navi: {}", value["data"]["navi"]);
-
-    //     match value["data"]["navi"] {
-    //         json::JsonValue::Array(ref vec) => Some(vec),
-    //         _ => None,
-    //     }
-
-    //     // match value["data"]["navi"] {
-    //     //     json::JsonValue::Null => None,
-    //     //     _ => Some(&value["data"]["navi"]),
-    //     // }
-
-    //     // match value["data"]["navi"] {
-    //     //     json::JsonValue::Object(ref object) => Some(object),
-    //     //     _ => {
-    //     //         // DBG
-    //     //         info!(
-    //     //             "fn navi cp2 navi: {} -- But not Object, return None.",
-    //     //             value["data"]["navi"]
-    //     //         );
-
-    //     //         None
-    //     //     }
-    //     // }
-    // }
-
-    // pub fn navi(&self) -> Option<&json::JsonValue> {
-    //     let data = self.data()?;
-    //     if value["data"]["navi"].is_empty() {
-    //         None
-    //     } else {
-    //         Some(&value["data"]["navi"])
-    //     }
-    // }
 
     pub fn subsections(&self) -> Option<&json::object::Object> {
         // let data = self.data()?;
